Changes as per code review suggestions.
