+++ iOS-Meme-Version1/Meme.swift
@@ -1,0 +1,17 @@
+//
+//  Meme.swift
+//  iOS-Meme-Version1
+//
+//  Created by Anupam Beri on 07/02/2021.
+//
+
+import Foundation
+import UIKit
+
+// MARK: The Meme Model object that inclues the text and images
+struct Meme {
+    let topText: String
+    let bottomText: String
+    let originalImage: UIImage
+    let memedImage: UIImage
+}
